Revert "Revert "feat: inbox request only required data""
This reverts commit 9577cb52641f63cc1cd59ee4d1473f25a26ec246.

Reason for revert: This revert wasn't needed as it was deemed that it wasn't the root cause of the issue Daniel found. We are putting this code back that was previously reviewed and tested.

refs VICE-4181
flag=react_inbox

Change-Id: I3790b89ebb5d2f84fc3770f98ba80c04a1358b39
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/345742
Reviewed-by: Balint Kartyas <balint.kartyas@instructure.com>
Reviewed-by: Roland Beres <roland.beres@instructure.com>
QA-Review: Roland Beres <roland.beres@instructure.com>
Product-Review: Omar Soto-Fortuño <omar.soto@instructure.com>
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>