spec: Remove unused binding_to_edit variable
Because this variable is never called explicitly, it is never created
(let statements are lazily evaluated).

It doesn't need to be here because a default account binding is created
when a dev key is first created, so tests pass without it.

I'm committing this separately so that the commits on top of it don't
have this miscellaneous change in them.

refs INTEROP-8575

Change-Id: Ie4fa88f666bdbaacba4bd202a79fdb1bdb817edf
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/347244
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Xander Moffatt <xmoffatt@instructure.com>
QA-Review: Tucker Mcknight <tmcknight@instructure.com>
Product-Review: Tucker Mcknight <tmcknight@instructure.com>