re-implement topnav in user settings
A previous commit, g/349439, we had to revert
the implementation of the TopNav in the user’s account
settings page (/profile/settings).

The previouis implementation was using
document.querySelector on a non-existent DOM element
which caused the page to break when the instui_topnav
FF was turned off (because mountPoint was not found
in the DOM).

At the time of this implmentation, IIRC, we did not have
initializeTopNavPortal() which now handles cases where
the mountPoint does not exist in the DOM.

This commit re-impliments the TopNav on the user’s account
settings page using initializeTopNavPortal().

This commit also updates the breadcrumbs to have two parts
(instead of just one part) regardless of the feature flag
setting. This change makes the breadcrumbs more consistent
with the rest of the links in the user’s settings sidebar.
Additionally, having two breadcrumbs avoids a bug with the
current version of the InstUI TopNavBar as documented here:

https://instructure.slack.com/archives/C0JCJ63TR/p1717523719748679

refs FOO-4085
flag=instui_nav

test plan:
- in new tab, navigate to User > Settings
  /profile/settings
- TopNav should be displayed with breadcrumbs
- breadcrumbs should have two parts:
  1. user short name linked to /profile route
  2. string “Settings”
- enable the instui-nav flag
- check >= tablet viewports
- page should fully render
- no errors in console
- disable instui-nav flag
- check >= tablet viewports
- reload /profile/settings
- page should fully render
- no errors in console

Change-Id: I232687f7debf2ed19ee580a881a44d9995dac484
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/349458
Reviewed-by: Martin Ochoa <martin.ochoa@instructure.com>
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
QA-Review: Michael Hulse <michael.hulse@instructure.com>
Product-Review: Michael Hulse <michael.hulse@instructure.com>