improve a11y of hamburger icon in TopNavBar
Note that RTL Jest tests are TBD for this commit as
the existing tests for TopNav.test.tsx (and related)
need to be improved to handle multiple viewport widths
and, in this case, testing the hamburger menu.

We’ll have to circle back around to our RTL tests in
order to add tests that cover the changes in this commit.

closes FOO-4330
refs INSTUI-4018
flag=instui_nav

test plan:
- log in to Canvas as an admin
- enable the “New InstUI navbar” feature flag
- refresh page
- navigate to Admin / Courses
- TopNavBar shoud render with hamburger menu
- if TopNavBar is expeanded, it will have:
    1. aria-label="Collapse menu"
    2. aria-expanded="true"
- if TopNavBar is collapsed, it will have:
    1. aria-label="Expand menu"
    2. aria-expanded="false"
- page should fully render
- no errors in console

Change-Id: I9aa72e943437c8d9443f621e5551bce56b1ac463
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/349589
QA-Review: Michael Hulse <michael.hulse@instructure.com>
Product-Review: Michael Hulse <michael.hulse@instructure.com>
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Martin Ochoa <martin.ochoa@instructure.com>