LIME Pagination Summary
This commit adds a pagination summary for the current page. It displays
the range of items currently being displayed and the total number of
items in the table.

test plan:

Load the Manage page and verify that the pagination summary is displayed
at the bottom of the table. Change the page number and verify that the
summary updates to reflect the new range of items being displayed.

flag=lti_registrations_page
closes INTEROP-8706

Change-Id: I71c6f1b3867254156d3366dcacc487aad4b8430d
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/351774
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Tucker Mcknight <tmcknight@instructure.com>
QA-Review: Tucker Mcknight <tmcknight@instructure.com>
Product-Review: Alexis Nast <alexis.nast@instructure.com>