Store Dynamic Registration URL
This commit adds a new column to the lti_ims_registrations table to
store the registration URL for the registration.
This is used to store the URL that was used to register the tool for
historical context, and so that it can be used to update the
registration in the future.

Test plan:

1. Install a tool using the dynamic registration flow
2. Run:
```
Lti::IMS::Registration.last
```
in a rails console and verify that the registration_url field is
populated

closes INTEROP-8735
flag=none

Change-Id: Ic55f665d88e9bce76f56b0e48c89d75e55e5fcf0
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/352798
Reviewed-by: Xander Moffatt <xmoffatt@instructure.com>
QA-Review: Xander Moffatt <xmoffatt@instructure.com>
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Migration-Review: Jake Oeding <jake.oeding@instructure.com>
Product-Review: Paul Gray <paul.gray@instructure.com>