Show warning on saving LTI 1.3 devkeys with resource_selection
Resource_selection is a placement that once was the default
for 1.1 tools, and was one of the very first placements for Canvas.
It makes the tool show up both in - assignment_selection
and link_selection (new module item), but is deprecated now.
We want to soft deprecate it for 1.3 tools by showing a warning message
when saving a devkey with resource_selection placement in it.

closes INTEROP-8782
flag=none

test plan:
- Save an 1.3 LTI dev key with resource_selection placement
- You should see a warning message

Change-Id: Id59b75f4f9738d5d80caa89455bf1e44d382f37c
Reviewed-on: https://gerrit.instructure.com/c/canvas-lms/+/354576
Tested-by: Service Cloud Jenkins <svc.cloudjenkins@instructure.com>
Reviewed-by: Csaba Csuzdi <csaba.csuzdi@instructure.com>
QA-Review: Csaba Csuzdi <csaba.csuzdi@instructure.com>
Product-Review: Alexis Nast <alexis.nast@instructure.com>